feat(config): remove redundant config codes
